Clean up repo, refactor code, merge psty_payload and psty_payload_large, fix examples accordingly
