+++ popsicle/deltas.txt
@@ -1,0 +1,1 @@
+{"29":{"ModN":{"q":29,"ds":[1,10,8,20,13,14,26,11,13,0,20,0,17,14,14,16,9,16,25,6,3,27,27,10,23]}},"47":{"ModN":{"q":47,"ds":[1,27,36,16,11,30,32,18,43,27,27,13,14,31,33,0,17,8,22,15,19]}},"23":{"ModN":{"q":23,"ds":[1,18,10,10,16,4,20,21,15,20,18,14,19,18,21,18,4,10,22,4,0,8,20,15,19]}},"19":{"ModN":{"q":19,"ds":[1,14,2,1,17,13,13,17,5,8,10,16,2,0,10,10,11,15,13,9,12,3,15,18,10]}},"37":{"ModN":{"q":37,"ds":[1,2,6,12,35,31,34,26,16,29,26,5,15,25,19,16,36,23,31,15,17]}},"2":{"Mod2":{"val":{"block":148920403746380739944295112161049118327}}},"13":{"ModN":{"q":13,"ds":[1,10,4,4,0,10,10,12,11,5,1,10,1,7,1,9,3,1,1,7,1,1,5,6,2,4,8,11,2,0,5,5]}},"43":{"ModN":{"q":43,"ds":[1,7,41,37,40,42,12,3,32,2,25,40,14,26,1,6,14,27,0,16,13]}},"11":{"ModN":{"q":11,"ds":[1,0,1,10,6,2,3,5,4,0,8,7,3,7,3,3,10,9,9,6,9,6,3,4,5,3,3,0,1,2,4,5]}},"17":{"ModN":{"q":17,"ds":[1,15,8,3,3,7,13,7,6,16,3,8,0,1,15,2,12,13,8,5,14,1,5,3,12]}},"31":{"ModN":{"q":31,"ds":[1,14,5,13,20,23,12,25,19,15,29,16,3,19,8,12,23,23,17,19,0,30,0,27,11]}},"41":{"ModN":{"q":41,"ds":[1,10,9,38,30,25,21,2,8,7,40,5,34,38,28,5,16,16,31,20,28]}},"3":{"Mod3":{"lsb":72666747907486993,"msb":17317893284980080846}},"5":{"ModN":{"q":5,"ds":[1,0,3,2,1,1,2,1,0,2,2,2,3,1,1,1,2,4,3,1,4,0,2,0,3,1,1,0,1,4,3,0,3,4,3,0,2,4,0,3,0,2]}},"7":{"ModN":{"q":7,"ds":[1,2,0,6,3,3,0,0,5,6,1,5,3,6,6,5,0,3,4,4,2,4,5,5,3,2,0,6,4,6,2,6,3,2,3,0,0,4,2,3,1,5]}},"53":{"ModN":{"q":53,"ds":[1,32,10,21,10,30,2,23,1,27,50,41,33,22,52,27,35,0,14,24,36]}}}
