+++ popsicle/examples/match-compute-parallel/client/run_client.sh
@@ -1,0 +1,16 @@
+echo "\n"
+echo "*******************Starting Files Preparation*******************"
+echo "\n"
+cargo run --example prepare_files_receiver  --features="psty_payload"
+wait
+echo "\n"
+echo "*******************Starting Computation Threads*******************"
+echo "\n"
+cargo run --example receiver_thread  --features="psty_payload" 0 &
+cargo run --example receiver_thread  --features="psty_payload" 1 &
+cargo run --example receiver_thread  --features="psty_payload" 2 &
+cargo run --example receiver_thread  --features="psty_payload" 3 &
+wait
+echo "*******************Starting Joining Threads Results*******************"
+echo "\n"
+cargo run --example join_aggregates_receiver --features="psty_payload" 4
