+++ popsicle/examples/match-compute-parallel/server/server_thread.rs
@@ -1,4 +1,4 @@
-use popsicle::psty_payload_large::{Sender};
+use popsicle::psty_payload::{Sender};
 
 use scuttlebutt::{AesRng, Block, Block512, TcpChannel};
